fix incorrect clip

+++ dogsgan/scripts/wgan.py
@@ -10,7 +10,7 @@ import dogsgan.models.dogs as dogs
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='run dcgan training')
     parser.add_argument('--lr', type=float, default=1e-4 ,help='learning rate')
-    parser.add_argument('--clip', type=int, default=0.02, help='weight clip threshold')
+    parser.add_argument('--clip', type=float, default=0.02, help='weight clip threshold')
     parser.add_argument('--noise_dim', type=int, default=1024, help='noise dimension')
     parser.add_argument('--base_dim', type=int, default=128, help='base dimension')
     parser.add_argument('--batch_size', type=int, default=64, help='batch size')
